+++ src/main/java/ortus/boxlang/modules/esapi/bifs/DecodeForHTML.java
@@ -19,12 +19,15 @@ import org.owasp.esapi.Encoder;
 
 import ortus.boxlang.runtime.bifs.BIF;
 import ortus.boxlang.runtime.bifs.BoxBIF;
+import ortus.boxlang.runtime.bifs.BoxMember;
 import ortus.boxlang.runtime.context.IBoxContext;
 import ortus.boxlang.runtime.scopes.ArgumentsScope;
 import ortus.boxlang.runtime.scopes.Key;
 import ortus.boxlang.runtime.types.Argument;
+import ortus.boxlang.runtime.types.BoxLangType;
 
 @BoxBIF
+@BoxMember( type = BoxLangType.STRING )
 public class DecodeForHTML extends BIF {
 
 	/**