+++ src/test/java/ortus/boxlang/modules/esapi/bifs/SanitizeHTMLTest.java
@@ -2,40 +2,17 @@ package ortus.boxlang.modules.esapi.bifs;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import ortus.boxlang.runtime.BoxRuntime;
-import ortus.boxlang.runtime.context.IBoxContext;
-import ortus.boxlang.runtime.context.ScriptingRequestBoxContext;
-import ortus.boxlang.runtime.scopes.IScope;
-import ortus.boxlang.runtime.scopes.Key;
-import ortus.boxlang.runtime.scopes.VariablesScope;
+import ortus.boxlang.modules.esapi.BaseIntegrationTest;
 
-public class SanitizeHTMLTest {
-
-	static BoxRuntime	instance;
-	IBoxContext			context;
-	IScope				variables;
-	static Key			result	= new Key( "result" );
-
-	@BeforeAll
-	public static void setUp() {
-		instance = BoxRuntime.getInstance( true );
-	}
-
-	@BeforeEach
-	public void setupEach() {
-		context		= new ScriptingRequestBoxContext( instance.getRuntimeContext() );
-		variables	= context.getScopeNearby( VariablesScope.name );
-	}
+public class SanitizeHTMLTest extends BaseIntegrationTest {
 
 	@DisplayName( "It sanitize using all policies" )
 	@Test
 	public void testAllPolicies() {
-		instance.executeSource(
+		runtime.executeSource(
 		    """
 		    result = sanitizeHTML( '<p>a <strong>link</strong> <a href="https://www.example.com" onClick="doSomethingBad()">test</a></p>' )
 		          """,