+++ changelog.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Missing member methods several Box Modules use
+- New Esapi Service to take care of service loading or instantiation capabilities
+- Muted initialization of ESAPI logger
+- Real integration testing
+- Updated to latest ESAPI
+
 ## [1.2.0] - 2025-02-22
 
 ### Fixed